requirements.txt
# --- Train ---
deepspeed
hydra-core
pytorch-lightning
wandb

# --- Tune ---
optuna
ray[tune]

# --- Project ---
cupy-cuda113
imageio